requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.46.0
greenlet==3.0.1
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
kombu==5.3.4
lml==0.1.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mutagen==1.47.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
passlib==1.7.4
Pillow==10.1.0
prompt-toolkit==3.0.41
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
seaborn==0.13.0
six==1.16.0
SQLAlchemy==2.0.23
texttable==1.7.0
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
WTForms==3.1.1
